Module 3 data-science-module/Problem Sets/Problem Set 4 - Streamlit/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
numpy>=1.24.0
sqlite3
